Add minor changes to models, views, admin. Add test for models, views, admin
Added tests for cook, dish & dish_type views that check if list & detail page are correctly retreived for the logged-in user
Added test for admin panel that check if the "years of experience" field is correctly displayed in various cases.
Added test for models to check if the redefined str representation is returned as expected & whether a cook with "years of experience" field is created correctly.
Removed "LoginRequiredMixin" from HomePageView.
Removed verbose & verbose_pluarl attributes DishType